Изменен тип свойства

+++ SmartWeatherBot/Database/IUserRepository.cs
@@ -9,7 +9,7 @@ namespace SmartWeatherBot.Database
     public interface IUserRepository
     {
         Task<User> GetUser(int index);
-        Task<User> GetUserTelegram(int telegram_id);
+        Task<User> GetUserTelegram(long telegram_id);
         Task<User> InsertAsync(User value);
         Task<bool> UpdateAsync(User value);
         Task<bool> DeleteAsync(int index);
